[NAVIGATION] - Refactor pretty-code
