+++ app/src/main/java/com/example/jetpackcomposetuto/MainActivity.kt
@@ -17,17 +17,7 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.navigation.NavType
-import androidx.navigation.compose.NavHost
-import androidx.navigation.compose.composable
-import androidx.navigation.compose.rememberNavController
-import androidx.navigation.navArgument
-import com.example.jetpackcomposetuto.model.Routes
-import com.example.jetpackcomposetuto.navigation.Screen1
-import com.example.jetpackcomposetuto.navigation.Screen2
-import com.example.jetpackcomposetuto.navigation.Screen3
-import com.example.jetpackcomposetuto.navigation.Screen4
-import com.example.jetpackcomposetuto.navigation.Screen5
+import com.example.jetpackcomposetuto.navigation.NavigationSetup
 import com.example.jetpackcomposetuto.ui.theme.JetpackComposeTutoTheme
 
 class MainActivity : ComponentActivity() {
@@ -39,36 +29,6 @@ class MainActivity : ComponentActivity() {
                 Surface(
                     color = MaterialTheme.colorScheme.background
                 ) {
-                    val navigationController = rememberNavController()
-                    NavHost(
-                        navController = navigationController,
-                        startDestination = Routes.Pantalla1.route
-                    ) {
-                        composable(Routes.Pantalla1.route) { Screen1(navigationController) }
-                        composable(Routes.Pantalla2.route) { Screen2(navigationController) }
-                        composable(Routes.Pantalla3.route) { Screen3(navigationController) }
-                        composable(
-                            Routes.Pantalla4.route,
-                            arguments = listOf(navArgument("age") { type = NavType.IntType })
-                        ) { backStackEntry ->
-                            Screen4(
-                                navigationController = navigationController,
-                                age = backStackEntry.arguments?.getInt("age") ?: 0
-                            )
-                        }
-                        composable(
-                            Routes.Pantalla5.route,
-                            arguments = listOf(navArgument("name"
-                            ) { defaultValue = "default Value" })
-                        ) { backStackEntry ->
-                            Screen5(
-                                navigationController = navigationController,
-                                name = backStackEntry.arguments?.getString("name")
-                            )
-                        }
-
-                    }
-
                     /*
              ***UNCOMMENT THE COMPOSABLE FUNCTION TO RENDER IN APP ***
                                       * */
@@ -91,7 +51,7 @@ class MainActivity : ComponentActivity() {
                     //SuperHeroRecyclerViewSticky()
                     //ScaffoldExample()
                     //DrawerMenu()
-
+                    NavigationSetup()
                 }
             }
 
@@ -99,6 +59,7 @@ class MainActivity : ComponentActivity() {
         //var myText by remember { mutableStateOf("Martin") }
         //MyRadioButtonList()
         // MyText(myText) { myText = it }
+
     }
 }
